+++ cdborm/errors.py
@@ -15,3 +15,4 @@ class FieldValidationError(Exception):
 class CanNotOverwriteRelationVariable(Exception): pass
 class AlreadyAssigned(Exception): pass
 class BadValueType(Exception): pass
+class NoDbSelected(Exception): pass