+++ History.md
@@ -6,6 +6,7 @@
   * Fixed problem with infinite loop made by relations.
   * Init now expects objects for relation, not id's.
   * Fixed choosing wrong database in Model.all
+  * get, all and save no raises NoDatabaseSelected when database == None.
 
 0.1.0 / 2012-12-15
 ==================